refactor: move files before splitting up generator (#26)
* refactor: move files before splitting up generator
The next PR I am working on is splitting up the SchemaGenerator into smaller chunks so it is easier to read and to test. Before I do that I am setting up the new package structure that I think works best. Since all the changes are to internal files this is a non-breaking change

* undo indentations