feat: move the default handling of CompletableFuture to code
This will have no impact on existing clients. I just moved the handling of CompletableFutures into the code so that it will always happen. Right now, clients need to remember to still handle CompletableFutures if they override the  hook.

We now resolve the type from the hook first, which defaults to the same type. Then we check after if it is on of the default supported wrapped types which right now only include CompletableFuture, but will extend to Publisher when we support subscriptions