feat: do not cache list types (#202)
* feat: do not cache list types

To simplify logic and make the cache simplier, we will not cache any list type. This means that the TypeCache is truely just caching fully qualified types. Lists will still be generated but they will not hit the cache to get their graphql type until they request a class.

This will also help if we want to support adding Sets to the types as it not just becomes one location we have to update for all the valid list types

* fix: delete unused extension method