fix: properly cache GraphQL object/interfaces/union types (#291)
`graphql-java` requires all `GraphQLObjectType`, `GraphQLInterfaceType` and `GraphQLUnionType` types to have unique names within a schema. While generating the schemas based on the reflections we have to ensure that each one of those objects is build only once. Since underlying objects do not implement hashcode it leads to problems if we attempt to update the fields through some hooks (e.g. use `newBuilder(existingObject)`). In order to make sure we only build those objects once we implemented basic caching logic in our builders - when we start building the target object we put it under construction and after it is build we add it to the cache and mark it completed.

Unfortunately, existing logic was flawed - all objects were being put under construction but only `GraphQLObjectType`s generated when processing their `GraphQLInterfaceType` were marked as completed and added to the cache. This change fixes the caching logic to properly complete all `GraphQLObjectType`, `GraphQLInterfaceType` and `GraphQLUnionType` types and add them to the underlying cache.