fix: allow modifying GraphQLInterfaceType in the willAddGraphQLTypeToSchema hook (#293)
* fix: allow modifying GraphQLInterfaceType in the willAddGraphQLTypeToSchema hook

InterfaceBuilder attempts to create all the implementing `GraphQLObjectType`s after building an interface but before executing the `willAddGraphQLTypeToSchema` hook. This means that if we attempt to update the interface in that hook there could be some objects implementing that interface and referencing the old one. Since graphql-java schema objects don't implement hashcode we end up with type conflict of two interfaces being defined in the schema.

Fix is to use `GraphQLTypeReference` when referencing `GraphQLInterfaceType` from the `GraphQLObjectType` builder. Reference is just by name and they are resolved as the last step of building a schema. This allows us to modify the interface in the hook.

Resolves: https://github.com/ExpediaDotCom/graphql-kotlin/issues/281

* update test to verify hook updates interface description