fix: *generateGraphQLType hooks are now called on interfaces
Fixes https://github.com/ExpediaDotCom/graphql-kotlin/issues/294

Interfaces now go through the internal TypeBuilder method 'graphQLTypeOf' instead of 'objectFromReflection'. This means we can make 'objectFromReflection' private. But the type may be (if not always) wrapped in a GraphQLNonNull which means we have to cast it back to the interface type before adding it to the ObjectBuilder.