fix context web filter test
ContextWebFilter returns a mono publisher that will compute the final result only after chain is subscribed to. StepVerifier was attempting to verify the context on the un-completed chain which did not populate the reactor context yet. Fix is to manually capture the context in the publisher chain, complete the publisher chain and then verify the contents of the captured reactor context.