[spring-server] enable support for specifying instrumentation order
`graphql-java` allows for applying various `Instrumentations` to the GraphQL execution. If multilple instrumentations are to be applied they need to be wrapped in an instance of `ChainedInstrumentation` that has an explicit order of the instrumentations.

In order to simplify Spring configuration we should allow users to either create a single instrumentation bean (could be a chained one) or multiple instrumentation beans and then automatically wrap them in chained instrumentation if necessary. Unfortunately, by default `graphql-java` `Instrumentation` does not specify order in which it should be applied (as it is generally explicitly configured). In order to support proper ordering of the instrumentations we can simply check if they implement `Ordered` interface and use their order value OR default to order of `0` otherwise (unordered behavior).