[federation] skip validation of nested objects under construction
When generating federated GraphQL schema we attempt to validate the generated objects. This is problematic when we are dealing with nested relationships as at validation time, those objects are still being built so we only have access to their `GraphQLTypeReference`. Those references are replaced as the last step of schema generation lifecycle which means we won't be able to perform any validations on those objects when they are still being built. This PR disables the validation of `@provides` directive when it references objects that are still being built.