Throw exception on multi-level interface
Resolves https://github.com/ExpediaGroup/graphql-kotlin/issues/419

According to the spec, an interface can not implement another interface. This is not caught by our schema generator but instead just returns the first level interface as the type in the schema. We are now throwing an exception to help translate this understanding of GraphQL through the Kotlin code