[federation] federated type resolvers should have access to environment
When resolving the federated types we currently don't have access to GraphQLContext. GraphQLContext can contain some important information (e.g. related to auth) that might be needed for type resolution so the resolvers should be able to access it. This PR updates the `FederatedTypeResolver` interface to also accept `DataFetchingEnvironment` when batch resolving the types.

Resolves: https://github.com/ExpediaGroup/graphql-kotlin/issues/472