[chore] move dokka configuration to release profile (#481)
In order to release to nexus we need to publish Javadocs. Currently dokka only works for JDK8 so we had it explicitly gated behind a profile that would only be activated when run using Java 8. Since we are explicitly gating Travis release job to run on Java 8 only so we don't need this separate profile.