Remove DataFetcherExecutionPredicate (#478)
Resolves https://github.com/ExpediaGroup/graphql-kotlin/issues/432

Remove the predicate and all usages to simplify the execution code. We already have a way of specifying a custom DataFetcherFactory which requires that developers still pass in a predicate when instead they can just override all the behaviour in the predicate in the custom DataFetcher.

This is a breaking change so it will require a release of 2.0.0. We may also want to verify our own internal usage of this change in Expedia first with a 2.0.0 RC release. While it should be fine it will be good to verify that a major change like this doesn't drastically impact any existing users