Add unit tests for renaming input type fields (#501)
* Add unit tests for renaming input type fields

Currently we can only parse renamed input class fields if they include the Jackson annotation @JsonProperty. This is because when we get the java class for the argument it doesn
't have any information for jackson to know that the constructor field has been renamed. I tried looking into adding a custom deserializer but we can't dynamically have one for all classes. Until we can find a better solution this unit test serves as documentation for the only way to make this work for now

* Update documentation on GraphQLName

* Fix linter

* Update renaming-fields.md