Fix subscription caching logic (#515)
* Refactor subscription caching logic

Move the saving of subscriptions to a separate class so we can vefify the logic with unit tests and simplify the
ApolloSubscriptionProtocolHandler. This also exposed a bug that we were not saving the operation subscriptions to be stopped properly. This is now covered by the unit tests

* Clear cache of operations if no more left

Prioritize clearing memory over saving the small amount of operation we have to perform is the session stays open but there is no active operations