BREAKING CHANGE: Refactor schema generator functions to top level file functions
Move the functions to generate a graphql type and generate the schema to top level file functions. This makes these functions pure since they accept the state of the SchemaGenerator as an argument and cleans up the methods that we need to expose in the class. It does however still make these public in graphql-kotlin-schema-generator as we still need to be able to call them from graphql-kotlin-federation