BREAKING CHANGE: Refactor graphQLTypeOf to top level function (#535)
* BREAKING CHANGE: Refactor schema generator functions to top level file functions

Move the functions to generate a graphql type and generate the schema to top level file functions. This makes these functions pure since they accept the state of the SchemaGenerator as an argument and cleans up the methods that we need to expose in the class. It does however still make these public in graphql-kotlin-schema-generator as we still need to be able to call them from graphql-kotlin-federation

* Make generateGraphQLType internal

* Remove redundant javadoc comments

* Move internal files into correct packages

* Move generateSchema into the SchemaGenerator

* Undo copyright update

* Move generateSchema into SchemaGenerator

* Move addExtendedTypes back to FederatedSchemaGenerator

* Move the add extended types into generateSchema

Remove the uneeded extra method in FederatedSchemaGenerator in favor of overriding generateSchema with the additional behaviour of adding all the types annotated with the ExtendsDirective before generating the schema