[build] apply nexus-publish plugin to explicitly create sonatype staging
Currently Gradle publish task is not using staging repository ID when publishing to Sonatype Maven Central (see: https://github.com/gradle/gradle/issues/5711). Furthermore, Gradle publish does not automatically release artifacts from Sonatype staging directory and requires manual step to finalize the release (i.e. log in to Sonatype repository and release uploaded staged artifacts).

As a workaround, in order to automatically publish to Maven Central we need to apply:
* `nexus-publish-plugin` - to explicitly create Sonatype staging ID prior to publishing artifacts (needed by `nexus-staging-plugin` to work reliably)
* `nexus-staging-plugin` - to explicitly close and release libs triggering the automatic Maven Central release

NOTE: It appears that shortcomings of Gradle `maven-publish` plugin are due to their preference of JCenter repository. Based on the documentation it should be possible to configure JCenter to automatically publish staged artifacts and subsequently sync them with Maven Central. If we ever update to publish to JCenter instead we should be able to simplify the build and remove those additional plugins.