feat: open FunctionDataFetcher methods for extension
Open the internal methods to the protected scope so that we can more easily override specific methods to provide more features. Internally to Expedia Group we have the need to provide a CustomFunctionDataFetcher that overrides the get method. However we don't want to have to reimplement all the other methods as well that covert the classes and parse the names. Right now we are actually missing out on using the new feature to use List over array for input types. This could be fixed internally but it is already done in open source so we would like to take advantage of that