Allow additional types to be added with custom SchemaGenerator (#587)
* Refactor additional types

This is a patch change to allow additional types to more easily be extended in later PRs. As a side-effect of this change we actually simplify where types are all generated and the main generateSchema function becomes just a little more straight forward

* Add more unit tests

* Simplify generateAdditionalTypes logic

After we have fully generated queries,mutations, and subscriptions there are no more type references. We only have this issue when we were adding classes in the middle of generate so this actually reduces logic and branches in the library

* Remove new features for now

Do not make the function open for outside calling right now. Discussing further with team and we can add the new feature in a separate PR