[docs] example usage of Reactor Mono
With the removal of evaluation predicates from the `FunctionDataFetcher` we no longer have an option to automatically handle Reactor/RxJava monads. In order to enable monad support we need to configure custom `SchemaGeneratorHook` to generate valid schema as well as custom `FunctionDataFetcher` so it can correctly process monads at runtime. Ideally this sort of additional runtime logic should be handled by hooks/instrumentation but with the current implementation relying on `graphql-java` it is the simplest workaround to get it working.