[generator] Replace GraphQLContext annotation with interface (#610)
* BREAKING CHANGE: Change GraphQLContext from annotation to interface

Instead of using an annotation and having to remember to include it in function arguments, we can create a marker interface for the GraphQLContext which library users can implement on a class to indicate that this class is the GraphQLContext. This moves the declaration from the arguments to the class itself which I think makes it more clear.

This is obviously a large breaking change to fundamental part of our library. It requires heavy documentation changes and review to take place before the 2.0 release. Please review and provided feedback on the interface package path as well

* Change GraphQLContextFactory to return the interface

* Combine interface and default implementation into single file

* Revert deletion of annotation class

Undo the deletion of the old annotation and instead make these changes a minor feature addition

* Update context docs

* Update docs wording

* typos

* typo

* typos, minor wording changes

* typos, minor wording changes

* Remove context annotation

* Update the docs

* GraphQLContextFactory must accept a GraphQLContext type

* Change generic type of ContextWebFilter

Instead of having a generic of a GraphQLContextFactory, we can make the generic for the type required of the context factory

* Simplify kFunction unit tests

Co-authored-by: Shane Myrick <accounts@shanemyrick.com>
Co-authored-by: Robert Del Favero <rdelfaverojr@expedia.com>