[build] attempt to use default mavenJava publication for plugin
Our current build does not upload Gradle plugin jar artifact to Sonatype. Reverting back to the original publication logic - originally had an issue with mutliple artifacts sharing the coordinates, checking if that is still an issue. If nothing else works we might need to disable Sonatype publish of Gradle artifact for now.