[generator] built-in support for optional input
In the GraphQL world, input types can be optional which means that the client can either:

* Not specify a value at all
* Send null explicitly
* Send the non-null type

This is in contrast with the JVM world where objects can either have some specific value or don't have any value (i.e. are `null`). As a result, when using default serialization logic it is not possible to distinguish between missing/unspecified value and explicit `null` value.

This PR introduces new `OptionalInput` sealed class wrapper that when used for query arguments will be automatically deserialized to `Undefined` object if argument was not specified OR to `Defined` wrapper class when argument was specified (including explicit null).

Resolves: https://github.com/ExpediaGroup/graphql-kotlin/issues/783