[3.x.x][generator] Only add directives with valid locations (#924)
* Only add directives with valid locations

When generating a schema you can have an annotation on any part of the schema, assuming you haven't set the Target, so if you are restricting the directive locations to only specific places they may still be added to the schema and then fail generation as they are added to invalid locations. While this could be viewed as a customer error, this is more of an issue with graphql-kotlin because we can have input and output objects with the same directives and things like the KeyDirective from federation are only valid on output. So in that case we shouldn't fail if someone uses the type as input elsewhere, it will still not be marked as a federation type for the input

* Rename function name

* Add unit tests for coverage

Co-authored-by: Shane Myrick <accounts@shanemyrick.com>