[plugin] fix generation of self referencing types (#948)
* [plugin] fix generation of self referencing types

Fixes logic to generate the self-referencing types. We were caching both generated `TypeSpecs` (class implementation) and `ClassNames` (class references) but we were doing so in wrong order. We rely on `ClassNames` and its selection set to determine whether the underlying type was already generated but we populated `ClassName` AFTER generating the type. This meant that if a type had a self-reference it would trigger generation of outer type first and then attempt to generate its own self reference but since `ClassName` was not yet generated it would generate the child type with the same name and then overwrite it with the generated parent. This lead to incorrect classes being generated as shown in original (invalid) `verify we can generate nested objects` test (note: that single type was generated even though child has different selection set). As a side effect it was also not possible to generate self-referencing input objects as such attempt would lead to infinite recursion and stack overflow.

Updated logic to first generate (and cache) target `ClassName`, store its selection set and then generate (and cache) the underlying `TypeSpec`. By storing the selection set of the generated `ClassName` we can also simplify the logic that compares different selection sets.

Resolves: https://github.com/ExpediaGroup/graphql-kotlin/issues/941

* update readme with latest changes