[server] add support for flow subscriptions (#972)
* [server] add support for flow subscriptions

`FlowSubscriptionExecutionStrategy` was available in the schema-generator library but we never integrated it with the spring-server. Since all our subscription handlers are based on the publisher API, I also updated flow subscription strategy to return publisher as well. This is a breaking for the existing flow subscription strategy users but without this change we would need to rewrite all our subscription handling logic to support flow natively (which we should do eventually when/if we move away from `graphql-java` based execution).

Explicit casts are required due to the type erasure and also Java reliance on site-variance vs Kotlin declaration-variance.

* update docs