[plugin] new generate-sdl GraphQL MOJO
Create new `generate-sdl` Maven goal that can generate GraphQL schema (in SDL format) as a build time artifact.

In order to genrate SDL in a generic way, we rely on the Java [Service Loader](https://docs.oracle.com/en/java/javase/11/docs/api/java.base/java/util/ServiceLoader.html) mechanism to dynamically locate an instance of a `SchemaGeneratorHooksProvider` that is responsible for creating an instance of the generate hooks that will be used to generate the schema. If no provider is found on the classpath, plugin will default to `NoopSchemaGeneratorHooks`.

MOJO requires list of packages that can contain GraphQL types and can be configured as shown below.

```xml
<plugin>
    <groupId>com.expediagroup</groupId>
    <artifactId>graphql-kotlin-maven-plugin</artifactId>
    <version>@graphql-kotlin.version@</version>
    <executions>
        <execution>
            <goals>
                <goal>generate-sdl</goal>
            </goals>
            <configuration>
                <packages>
                    <package>com.example</package>
                </packages>
            </configuration>
        </execution>
    </executions>
    <!-- optional service provider dependency -->
    <dependencies>
        <dependency>
            <groupId>com.expediagroup</groupId>
            <artifactId>graphql-kotlin-federated-hooks-provider</artifactId>
            <version>${graphql-kotlin.version}</version>
        </dependency>
    </dependencies>
</plugin>
```