Generate DataLoaders on every request (#1040)
With the introduction of the new KotlinDataLoader interface we made the way to get the data loaders a property. However taking the advice from graphql-java we should generate each DataLoader every request. The factory generates the registry for every request currently but the data loaders themselves would still be static.

This also updates the docs and examples around data loaders from the last PR