Support graphql-java `GraphQLContext` injection
With graphql-java 17.x injecting a custom context object was deprecated.
The new preferred context method is to use a mutable key-value map defined by graphql-java.

As the marker interface in graphql-kotlin can not be applied to `GraphQLContext` in
graphql-java, downstream projects can not migrate to the new context object.

This change makes it possible to use graphql-java `GraphQLContext` objects like custom
objects. The marker interface in graphql-kotlin is deprecated but not removed to offer
a migration path.

See also:
- https://github.com/graphql-java/graphql-java/pull/2368