Custom types with annotation (#1298)
### :pencil: Description
This adds a new annotation `@GraphQLType` which you can use to specify the return type of any field to be different than the Kotlin type. It explicitly takes a string for the type name and not just another `KClass` reference, because what we can do is return the field type as a type reference, but then allow devs to specify custom additional types that they have built up with their own GraphQLType builders.

The prime use case for this is what we were trying to accomplish with `@GraphQLUnion`. In the scenario where you want to define a new union that is many different Kotlin classes but maybe those come from other libraries and you can't modify the class to have it implement the empty interface, this can solve that solution. But now you can also specify directives, descriptions, and deprecations on the type. Plus this solution scales beyond just unions and can work for any custom type that you can define in Kotlin code. 

If we feel like this is a decent approach, I would even consider deprecating the `@GraphQLUnion` annotation in favor of this generic approach that anyone can use and even extend now that we have added the ability to pass in `GraphQLType`s in the config.

### :link: Related Issues