[generator] add runtime warning on deprecated context usage (#1323)
### :pencil: Description
This PR updates `FunctionDataFetcher` so that it logs a warning when injecting a `GraphQLContext`.  This functionality is [deprecated](https://github.com/ExpediaGroup/graphql-kotlin/blob/master/website/docs/schema-generator/execution/contextual-data.md#interface-injection-deprecated).  Unlike the `GraphQLContext` interface itself, which is annotated `@deprecated, there is no compile or runtime time warning when using `GraphQLContext` interface injection in a query or mutation.

### :link: Related Issues
* discussion: #1305
* supersedes: #1306

Co-authored-by: Joel Brown <jbrown@classpass.com>