[generator] filter lambda properties
Currently we don't support lambda parameters. Attempting to use a lambda property blows up with somewhat ambiguous messages (lambda return type is outside of supported packages/cannot calculate erasure type for suspendable lambda). Currently it was also not possible to filter out lambda properties using `@GraphQLIgnore` annotations as it was applied AFTER attempting to read lambda erasure type (which blows up as it cannot be calculated).

This PR adds new property filter that checks whether property is a lambda and throws `InvalidPropertyReturnTypeException` if it is. Also moved `@GraphQLIgnore` filter to be applied before we attempt to calculate JVM erasure (used to verify property type is not blacklisted).

Throwing exception from a filter might not be ideal but IMHO is a better alternative than throwing current ambiguous exception (about JVM erasure) or by just filtering lambda parameters (which leads to questions of why my public property is not available in the graph).