[generator] add IDValueUnboxer for correctly serializing ID value class
Since `ID` is a value class, it may be represented at runtime as a wrapper or directly as underlying type. Due to the generic
nature of the query processing logic we *always* end up with up a wrapper type when resolving the field value. As a result,
in order to ensure that underlying scalar value is correctly serialized, we need to explicitly unwrap it by registering
`IDValueUnboxer` with your GraphQL instance.

```kotlin
// registering custom value unboxer
val graphQL = GraphQL.newGraphQL(graphQLSchema)
    .valueUnboxer(IDValueUnboxer())
    .build()
```

`IDValueUnboxer` is automatically registered by `graphql-kotlin-spring-server`.