Update spring-server to use FlowSubscriptionSchemaGeneratorHooks (#1479)
Currently the `spring-server` automatically configures the subscription strategy that supports Kotlin `Flow`  but does not configure the hooks neccessary to use it. Updating `SubscriptionAutoConfiguration` to create `FlowSubscriptionSchemaGeneratorHooks` if bean is not available.

Cleaned up the federation docs to highlight that both hooks and execution strategy are needed for using `Flow`s in subscriptions. Added a note that the subscription protocol implemented in `spring-server` is deprecated.