[ktor plugin] refactor route configuration
It appears that it is currently not possible to easily access the previously configured routes. This means that it is not possible to intercept our auto-configured routes to add some additional configuration (e.g. `Authentication`). This PR refactors the plugin to extract all routes into a separate extension functions that can be invoked by the users.

New plugin configuration will consist of 2 steps - (1) configuring server and (2) configuring routes.

```kotlin
fun Application.myGraphQLModule() {
    install(GraphQL) {
        // all schema, engine and server configuration go here
    }
    install(Routing) {
        // configure routes
        graphQLGetRoute()
        graphQLPostRoute()
        graphQLSDLRoute()
        graphiQLRoute()
    }
}
```

This refactor follows the same pattern as `Authentication` plugin.

Resolves: https://github.com/ExpediaGroup/graphql-kotlin/issues/1676