feat(batching): avoid calculating document ast height in advance (#1800)
Batching by LEVEL requires to calculate the height of all graphql
documents in a batch request to be calculated before deciding of a
certain level was dispatched, the reason is because a batch request
could contain multiple and potentially different operations, with
different levels, so, to decide if a level was dispatched, we need to
know which operations in the batch request share the same level.

This PR will remove that logic, which was take it from the graphql-java
[MaxDepthInstrumentation](https://github.com/graphql-java/graphql-java/blob/master/src/main/java/graphql/analysis/MaxQueryDepthInstrumentation.java#L56).

Instead we will calculate the height on runtime, simplifying the
dispatch by level instrumentation and removing some CPU usage by not
doing an AST traversal to calculate height.