feat(generator): add didGenerateDirective hook (#1818)
### :pencil: Description

Adds new `didGenerateDirective` hook that is invoked AFTER generating
the directive. This PR also fix directive generation logic to ensure
hooks are only called once (when generating the directive). Previously
we were always invoking `willGenerateDirective` hook regardless whether
the directive was already generated or not (which lead to creation of
new directive when modifying the directive in the new
`didGenerateDirective` hook).

### :link: Related Issues

Type generation still suffers from the same bug (#1815)