fix(federation): skip fieldset validation when it includes type reference
We are performing federated directive validations when generating GraphQL schema. This is an extra convenience as it gives us information about invalid schema sooner. Subgraph schemas will then be re-validated during the supergraph composition logic.

Currently it is possible to create `@key` that references the type still under construction (`GraphQLReferenceType`). Since we cannot validate fields on the type that doesn't exist yet, we cannot validate it either. Updating logic to log a warning when this happens and skip validation logic.