feat: graphql java 23 update (#2106)
### :pencil: Description
https://github.com/graphql-java/graphql-java/releases/tag/v23.1
https://github.com/graphql-java/graphql-java/releases/tag/v23.0

graphql-java 22 introduced the support to `@defer` directive, however,
it wasn't added to the SDL, now in graphql-java 23 `@defer` is being,
added. For now, we will disable `@defer` because work will be needed to
implement the transport of responses through HTTP.

graphql-java 23 broke the data loader dispatch by synchronous execution
exhaustion mechanism, because of the introduction of dataLoader hooks,
the reason is because the mechanism was relying on the number of
dependents of the CompletableFuture returned by the dataLoader, which
now is instrumented and adds extra dependents.

significant changes were done to stop relying on number of dependents
and instead, use the instrumentation provided by the data loader library

A KotlinDataLoaderRegistry decorator is no longer needed, with this we
will simply the execution.

---------

Co-authored-by: Samuel Vazquez <samvazquez@expediagroup.com>